feat: allow defined object in `optional()`
